requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.2.0
importlib-resources==6.4.0
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
zipp==3.19.2
